moving function mg_sleep_and_assert to a common layer in mg_utils.py Using function in another test file
