+++ tests/e2e/replication/CMakeLists.txt
@@ -12,3 +12,4 @@ copy_e2e_python_files(replication_show show.py)
 copy_e2e_python_files(replication_show show_while_creating_invalid_state.py)
 copy_e2e_python_files_from_parent_folder(replication_show ".." memgraph.py)
 copy_e2e_python_files_from_parent_folder(replication_show ".." interactive_mg_runner.py)
+copy_e2e_python_files_from_parent_folder(replication_show ".." mg_utils.py)
