+++ tests/e2e/replication/show_while_creating_invalid_state.py
@@ -13,9 +13,9 @@ import sys
 
 import os
 import pytest
-import time
 
 from common import execute_and_fetch_all
+from mg_utils import mg_sleep_and_assert
 import interactive_mg_runner
 import mgclient
 import tempfile
@@ -59,21 +59,6 @@ MEMGRAPH_INSTANCES_DESCRIPTION = {
         ],
     },
 }
-
-
-def mg_sleep_and_assert(expected_value, function_to_retrieve_data, max_duration=20, time_between_attempt=0.5):
-    result = function_to_retrieve_data()
-    start_time = time.time()
-    while result != expected_value:
-        current_time = time.time()
-        duration = current_time - start_time
-        if duration > max_duration:
-            assert False, " mg_sleep_and_assert has tried for too long and did not get the expected result!"
-
-        time.sleep(time_between_attempt)
-        result = function_to_retrieve_data()
-
-    return result
 
 
 def test_show_replicas(connection):
